fix: premature guess submission on virtual keys
When working on #30 the previous fix implemented in #24 was broken again, this fixes the issue one more time.

The thing is that by moving the virtual keyboard as a web component, the document.activeElement only returns the root shadow node, so validating against the class name to be "virtual-key" is not suitable anymore.